backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
pydantic>=2.6.4
pyjwt>=2.10.1
bcrypt==4.1.3
passlib>=1.7.4
python-jose==3.5.0
python-multipart>=0.0.9
email-validator>=2.2.0
requests>=2.31.0
tzdata>=2024.2
